third_party/WebKit/LayoutTests/inspector/sources/debugger-breakpoints/debugger-reload-breakpoints-with-source-maps-expected.txt

Tests "reload" from within inspector window while on pause.


Breakpoint sidebar pane before reload:
source1.js:15    var handler = new ClickHandler();
source1.js:17}
Page reloaded.
Breakpoint sidebar pane after reload:
source1.js:15    var handler = new ClickHandler();
source1.js:17}
